Cognitive Modelling Assignment Masters Module UCD
In the training phase of this experiment, participants were given the
16 training items (see excel document) from which they learned to
identify diseases.
Participants studied these 16 training items and were assessed to
ensure they had learned the disease categories successfully.
In the test phase, participants were repeatedly shown 5 test items (see
Table 2 in excel sheet).  Each time they were shown a given test item,
they were asked to rate that item as a member of a different category…

(category A, category B, and category C) or a different conjunction
(conjunctions A&B, A&C, and B&C).
Participants rated membership on a scale going from -10 (definitely
non-member), to 0 (neutral) to +10 (definitely member).
The task was to come up with a computational model that, for each test
item (test items 1 to 5), computes that item's degree of membership in
the three categories A, B and C.
The computational model was then extended to compute each test item's
degree of membership in the three conjunctions A&B, A&C, and B&C.

A presentation of my model and comparative essay of other models
presented in class is also attached.